refactor: Move gRPC make_server up a module
