+++ src/influxdb_ioxd/rpc.rs
@@ -13,3 +13,44 @@ pub mod expr;
 pub mod id;
 pub mod input;
 pub mod service;
+
+use arrow_deps::arrow_flight::flight_service_server::FlightServiceServer;
+use data_types::error::ErrorLogger;
+use generated_types::{i_ox_testing_server::IOxTestingServer, storage_server::StorageServer};
+use query::DatabaseStore;
+use snafu::{ResultExt, Snafu};
+use std::sync::Arc;
+use tokio::net::TcpListener;
+use tokio_stream::wrappers::TcpListenerStream;
+
+#[derive(Debug, Snafu)]
+pub enum Error {
+    #[snafu(display("gRPC server error:  {}", source))]
+    ServerError { source: tonic::transport::Error },
+}
+
+pub type Result<T, E = Error> = std::result::Result<T, E>;
+
+/// Instantiate a server listening on the specified address
+/// implementing the IOx, Storage, and Flight gRPC interfaces, the
+/// underlying hyper server instance. Resolves when the server has
+/// shutdown.
+pub async fn make_server<T>(socket: TcpListener, storage: Arc<T>) -> Result<()>
+where
+    T: DatabaseStore + 'static,
+{
+    let stream = TcpListenerStream::new(socket);
+
+    tonic::transport::Server::builder()
+        .add_service(IOxTestingServer::new(service::GrpcService::new(
+            storage.clone(),
+        )))
+        .add_service(StorageServer::new(service::GrpcService::new(
+            storage.clone(),
+        )))
+        .add_service(FlightServiceServer::new(service::GrpcService::new(storage)))
+        .serve_with_incoming(stream)
+        .await
+        .context(ServerError {})
+        .log_if_error("Running Tonic Server")
+}
